requirements.txt
streamlit==1.41.1
tensorflow==2.18.0
numpy==2.2.1
Pillow==11.0.0